+++ tests/CfdiUtilsTests/Utils/Internal/TemporaryFileTest.php
@@ -2,7 +2,7 @@
 namespace CfdiUtilsTests\Utils\Internal;
 
 use CfdiUtils\Utils\Internal\TemporaryFile;
-use PHPUnit\Framework\TestCase;
+use CfdiUtilsTests\TestCase;
 
 class TemporaryFileTest extends TestCase
 {
@@ -26,7 +26,7 @@ class TemporaryFileTest extends TestCase
         }
     }
 
-    public function testCreateWouldFailIfCannotCreateTheFile()
+    public function testCreateOnNonExistentFolderThrowsException()
     {
         $directory = __DIR__ . '/non/existent/directory/';
         $this->expectException(\RuntimeException::class);
@@ -34,18 +34,80 @@ class TemporaryFileTest extends TestCase
         TemporaryFile::create($directory);
     }
 
-    public function testCreateOnReadOnlyFolder()
+    public function testCreateOnReadOnlyFolderThrowsException()
     {
+        // redirect test to MS Windows case
+        if ($this->isRunningOnWindows()) {
+            // don't know how to test this on AppVeyor
+            // since MS Windows create file on folder with chmod 0400
+            // and AppVeyor allow write on WINDIR
+            $this->markTestSkipped('Cannot create scenario to perform this test on MS Windows');
+            return;
+        }
+
+        // prepare directory
         $directory = __DIR__ . '/readonly';
         mkdir($directory);
-        chmod($directory, 0550);
+        chmod($directory, 0500);
+
+        // setup exception
         $this->expectException(\RuntimeException::class);
         $this->expectExceptionMessage('Unable to create a temporary file');
         try {
             TemporaryFile::create($directory);
         } finally {
-            chmod($directory, 0770);
+            // cleanup
+            chmod($directory, 0700);
             rmdir($directory);
         }
+    }
+
+    public function testObjectToStringBehavior()
+    {
+        $file = TemporaryFile::create();
+        $this->assertSame($file->getPath(), (string) $file);
+        $file->remove();
+    }
+
+    public function testReadAndWrite()
+    {
+        $content = 'Lorem Ipsum';
+        $file = TemporaryFile::create();
+        $this->assertSame('', $file->retriveContents(), 'Contents should be empty');
+
+        $file->storeContents($content);
+        $this->assertSame($content, $file->retriveContents(), 'Contents should be what we have store');
+
+        $file->remove();
+    }
+
+    public function testRunAndRemoveGreenPath()
+    {
+        $file = TemporaryFile::create();
+        $expected = 'foo';
+
+        $retrieved = $file->runAndRemove(function () use ($expected) {
+            return $expected;
+        });
+
+        $this->assertSame($expected, $retrieved, 'Method did not return the expected value');
+        $this->assertFileNotExists($file);
+    }
+
+    public function testRunAndRemoveWithException()
+    {
+        $file = TemporaryFile::create();
+
+        try {
+            $file->runAndRemove(function () {
+                throw new \RuntimeException('DUMMY');
+            });
+        } catch (\RuntimeException $exception) {
+            if ('DUMMY' !== $exception->getMessage()) {
+                throw new \RuntimeException('Expected exception was not thrown', 0, $exception);
+            }
+        }
+
+        $this->assertFileNotExists($file);
     }
 }